requirements.txt
streamlit==0.76.0
tabula-py==2.2.0
pandas==1.1.5
XlsxWriter==1.2.9
